split api_testrail - step 1
